+++ ServiceProviderBot/Bot/Dialogs/Feedback/FeedbackDialog.cs
@@ -1,9 +1,12 @@
-﻿using Microsoft.Bot.Builder.Dialogs;
+﻿using Microsoft.Bot.Builder;
+using Microsoft.Bot.Builder.Dialogs;
 using Microsoft.Extensions.Configuration;
 using ServiceProviderBot.Bot.State;
 using Shared;
 using Shared.ApiInterface;
 using Shared.Prompts;
+using System.Threading;
+using System.Threading.Tasks;
 
 namespace ServiceProviderBot.Bot.Dialogs.Feedback
 {
@@ -14,37 +17,39 @@ namespace ServiceProviderBot.Bot.Dialogs.Feedback
         public FeedbackDialog(StateAccessors state, DialogSet dialogs, IApiInterface api, IConfiguration configuration)
             : base(state, dialogs, api, configuration) { }
 
-        public override WaterfallDialog GetWaterfallDialog()
+        public override Task<WaterfallDialog> GetWaterfallDialog(ITurnContext turnContext, CancellationToken cancellation)
         {
-            // Define the dialog and add it to the set.
-            return new WaterfallDialog(Name, new WaterfallStep[]
+            return Task.Run(() =>
             {
-                async (dialogContext, cancellationToken) =>
+                return new WaterfallDialog(Name, new WaterfallStep[]
                 {
-                    // Prompt for feedback.
-                    return await dialogContext.PromptAsync(
-                        Prompt.TextPrompt,
-                        new PromptOptions {
-                            Prompt = Phrases.Feedback.GetFeedback
-                        },
-                        cancellationToken);
-                },
-                async (dialogContext, cancellationToken) =>
-                {
-                    var user = await this.api.GetUser(dialogContext.Context);
+                    async (dialogContext, cancellationToken) =>
+                    {
+                        // Prompt for feedback.
+                        return await dialogContext.PromptAsync(
+                            Prompt.TextPrompt,
+                            new PromptOptions {
+                                Prompt = Phrases.Feedback.GetFeedback
+                            },
+                            cancellationToken);
+                    },
+                    async (dialogContext, cancellationToken) =>
+                    {
+                        var user = await this.api.GetUser(dialogContext.Context);
 
-                    // Save the feedback.
-                    var feedback = new EntityModel.Feedback();
-                    feedback.SenderId = user.Id;
-                    feedback.Text = (string)dialogContext.Result;
-                    await this.api.Create(feedback);
+                        // Save the feedback.
+                        var feedback = new EntityModel.Feedback();
+                        feedback.SenderId = user.Id;
+                        feedback.Text = (string)dialogContext.Result;
+                        await this.api.Create(feedback);
 
-                    // Send thanks.
-                    await Messages.SendAsync(Phrases.Feedback.Thanks, dialogContext.Context, cancellationToken);
+                        // Send thanks.
+                        await Messages.SendAsync(Phrases.Feedback.Thanks, dialogContext.Context, cancellationToken);
 
-                    // End this dialog to pop it off the stack.
-                    return await dialogContext.EndDialogAsync(cancellationToken);
-                }
+                        // End this dialog to pop it off the stack.
+                        return await dialogContext.EndDialogAsync(cancellationToken);
+                    }
+                });
             });
         }
     }
