+++ ServiceProviderBot/Bot/Dialogs/MasterDialog.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.Bot.Builder.Dialogs;
+﻿using Microsoft.Bot.Builder;
+using Microsoft.Bot.Builder.Dialogs;
 using Microsoft.Extensions.Configuration;
 using ServiceProviderBot.Bot.Dialogs.Capacity;
 using ServiceProviderBot.Bot.Dialogs.Feedback;
@@ -8,6 +9,8 @@ using Shared.ApiInterface;
 using Shared.Prompts;
 using System;
 using System.Linq;
+using System.Threading;
+using System.Threading.Tasks;
 
 namespace ServiceProviderBot.Bot.Dialogs
 {
@@ -18,99 +21,102 @@ namespace ServiceProviderBot.Bot.Dialogs
         public MasterDialog(StateAccessors state, DialogSet dialogs, IApiInterface api, IConfiguration configuration)
             : base(state, dialogs, api, configuration) { }
 
-        public override WaterfallDialog GetWaterfallDialog()
+        public override Task<WaterfallDialog> GetWaterfallDialog(ITurnContext turnContext, CancellationToken cancellation)
         {
-            return new WaterfallDialog(Name, new WaterfallStep[]
+            return Task.Run(() =>
             {
-                async (dialogContext, cancellationToken) =>
+                return new WaterfallDialog(Name, new WaterfallStep[]
                 {
-                    // Check if the user is already registered.
-                    var user = await api.GetUser(dialogContext.Context);
-                    if (user == null)
+                    async (dialogContext, cancellationToken) =>
                     {
-                        await Messages.SendAsync(Phrases.Greeting.NotRegistered(dialogContext.Context), dialogContext.Context, cancellationToken);
-                        return await dialogContext.EndDialogAsync(cancellationToken);
-                    }
-
-                    // Check if we already have an organization for the user.
-                    var organization = await api.GetOrganization(user.OrganizationId);
-                    if (organization == null)
-                    {
-                        await Messages.SendAsync(Phrases.Greeting.NoOrganization, dialogContext.Context, cancellationToken);
-                        return await dialogContext.EndDialogAsync(cancellationToken);
-                    }
-
-                    // Check if the organization is verified.
-                    if (!organization.IsVerified)
-                    {
-                        await Messages.SendAsync(Phrases.Greeting.UnverifiedOrganization, dialogContext.Context, cancellationToken);
-                        return await dialogContext.EndDialogAsync(cancellationToken);
-                    }
-
-                    // Save the user ID and organization ID to the user context so that
-                    // they can be accessed by other dialogs without API lookups.
-                    var userContext = await this.state.GetUserContext(dialogContext.Context, cancellationToken);
-                    userContext.UserId = user.Id;
-                    userContext.OrganizationId = user.OrganizationId;
-
-                    // Check if the initial message is one of the keywords.
-                    var incomingMessage = dialogContext.Context.Activity.Text;
-                    if (!string.IsNullOrEmpty(incomingMessage))
-                    {
-                        bool isKeyword = Phrases.Keywords.List.Any(k => string.Equals(incomingMessage, k, StringComparison.OrdinalIgnoreCase));
-                        if (isKeyword)
-                        {
-                            return await dialogContext.NextAsync(incomingMessage);
-                        }
-                    }
-
-                    // Prompt for a keyword.
-                    return await dialogContext.PromptAsync(
-                        Prompt.GreetingTextPrompt,
-                        new PromptOptions {
-                            Prompt = Phrases.Greeting.GetKeywords(user, welcomeUser: true),
-                            RetryPrompt = Phrases.Greeting.GetKeywords(user)
-                        },
-                        cancellationToken);
-                },
-                async (dialogContext, cancellationToken) =>
-                {
-                    var result = dialogContext.Result as string;
-
-                    if (string.Equals(result, Phrases.Keywords.Update, StringComparison.OrdinalIgnoreCase))
-                    {
-                        // Push the update dialog onto the stack.
-                        return await BeginDialogAsync(dialogContext, UpdateCapacityDialog.Name, null, cancellationToken);
-                    }
-                    else if (string.Equals(result, Phrases.Keywords.Enable, StringComparison.OrdinalIgnoreCase) ||
-                             string.Equals(result, Phrases.Keywords.Disable, StringComparison.OrdinalIgnoreCase))
-                    {
-                        // Enable/disable contact.
-                        var enable = string.Equals(result, Phrases.Keywords.Enable, StringComparison.OrdinalIgnoreCase);
-
+                        // Check if the user is already registered.
                         var user = await api.GetUser(dialogContext.Context);
-                        if (user.ContactEnabled != enable)
+                        if (user == null)
                         {
-                            user.ContactEnabled = enable;
-                            await this.api.Update(user);
+                            await Messages.SendAsync(Phrases.Greeting.NotRegistered(dialogContext.Context), dialogContext.Context, cancellationToken);
+                            return await dialogContext.EndDialogAsync(cancellationToken);
                         }
 
-                        await Messages.SendAsync(Phrases.Greeting.ContactEnabledUpdated(user.ContactEnabled), dialogContext.Context, cancellationToken);
+                        // Check if we already have an organization for the user.
+                        var organization = await api.GetOrganization(user.OrganizationId);
+                        if (organization == null)
+                        {
+                            await Messages.SendAsync(Phrases.Greeting.NoOrganization, dialogContext.Context, cancellationToken);
+                            return await dialogContext.EndDialogAsync(cancellationToken);
+                        }
+
+                        // Check if the organization is verified.
+                        if (!organization.IsVerified)
+                        {
+                            await Messages.SendAsync(Phrases.Greeting.UnverifiedOrganization, dialogContext.Context, cancellationToken);
+                            return await dialogContext.EndDialogAsync(cancellationToken);
+                        }
+
+                        // Save the user ID and organization ID to the user context so that
+                        // they can be accessed by other dialogs without API lookups.
+                        var userContext = await this.state.GetUserContext(dialogContext.Context, cancellationToken);
+                        userContext.UserId = user.Id;
+                        userContext.OrganizationId = user.OrganizationId;
+
+                        // Check if the initial message is one of the keywords.
+                        var incomingMessage = dialogContext.Context.Activity.Text;
+                        if (!string.IsNullOrEmpty(incomingMessage))
+                        {
+                            bool isKeyword = Phrases.Keywords.List.Any(k => string.Equals(incomingMessage, k, StringComparison.OrdinalIgnoreCase));
+                            if (isKeyword)
+                            {
+                                return await dialogContext.NextAsync(incomingMessage);
+                            }
+                        }
+
+                        // Prompt for a keyword.
+                        return await dialogContext.PromptAsync(
+                            Prompt.GreetingTextPrompt,
+                            new PromptOptions {
+                                Prompt = Phrases.Greeting.GetKeywords(user, welcomeUser: true),
+                                RetryPrompt = Phrases.Greeting.GetKeywords(user)
+                            },
+                            cancellationToken);
+                    },
+                    async (dialogContext, cancellationToken) =>
+                    {
+                        var result = dialogContext.Result as string;
+
+                        if (string.Equals(result, Phrases.Keywords.Update, StringComparison.OrdinalIgnoreCase))
+                        {
+                            // Push the update dialog onto the stack.
+                            return await BeginDialogAsync(dialogContext, UpdateCapacityDialog.Name, null, cancellationToken);
+                        }
+                        else if (string.Equals(result, Phrases.Keywords.Enable, StringComparison.OrdinalIgnoreCase) ||
+                                 string.Equals(result, Phrases.Keywords.Disable, StringComparison.OrdinalIgnoreCase))
+                        {
+                            // Enable/disable contact.
+                            var enable = string.Equals(result, Phrases.Keywords.Enable, StringComparison.OrdinalIgnoreCase);
+
+                            var user = await api.GetUser(dialogContext.Context);
+                            if (user.ContactEnabled != enable)
+                            {
+                                user.ContactEnabled = enable;
+                                await this.api.Update(user);
+                            }
+
+                            await Messages.SendAsync(Phrases.Greeting.ContactEnabledUpdated(user.ContactEnabled), dialogContext.Context, cancellationToken);
+                            return await dialogContext.EndDialogAsync(cancellationToken);
+                        }
+                        else if (string.Equals(result, Phrases.Keywords.Feedback, StringComparison.OrdinalIgnoreCase))
+                        {
+                            // Push the feedback dialog onto the stack.
+                            return await BeginDialogAsync(dialogContext, FeedbackDialog.Name, null, cancellationToken);
+                        }
+
+                        return await dialogContext.NextAsync(cancellationToken);
+                    },
+                    async (dialogContext, cancellationToken) =>
+                    {
+                        // End this dialog to pop it off the stack.
                         return await dialogContext.EndDialogAsync(cancellationToken);
                     }
-                    else if (string.Equals(result, Phrases.Keywords.Feedback, StringComparison.OrdinalIgnoreCase))
-                    {
-                        // Push the feedback dialog onto the stack.
-                        return await BeginDialogAsync(dialogContext, FeedbackDialog.Name, null, cancellationToken);
-                    }
-
-                    return await dialogContext.NextAsync(cancellationToken);
-                },
-                async (dialogContext, cancellationToken) =>
-                {
-                    // End this dialog to pop it off the stack.
-                    return await dialogContext.EndDialogAsync(cancellationToken);
-                }
+                });
             });
         }
     }
